Fix a cpplint allocator regression. (#590)
We can't unconditionally include rcutils/allocator.h in the
msg__functions.c.em, because that file is used multiple times
when generating services.  That causes cpplint to complain
about a duplicate include.  Instead put the include in
idl__functions.c.em, and mark it as already included so we
won't try to include it again.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>